fix : fix go test in windows
